+++ shared/src/index.ts
@@ -11,3 +11,4 @@ export * from './donation-bank';
 export * from './donation-methods';
 export * from './donation-amounts';
 export * from './donation-payment-modes';
+export * from './video-embed';